SEMANA10/dia6/pos_backend_docker/app/requirements.txt
django==3.2
djangorestframework
mysqlclient
django-cors-headers
cloudinary
djangorestframework-simplejwt
drf-yasg
python-decouple
gunicorn